# services/users/cmd/docs/swagger.yaml
basePath: /
definitions:
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.AddressRequest:
    properties:
      address:
        type: string
      principal:
        type: boolean
    required:
    - address
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.Device:
    properties:
      device_name:
        type: string
      public_key:
        type: string
    required:
    - device_name
    - public_key
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.EmailRequest:
    properties:
      email:
        type: string
      principal:
        type: boolean
    required:
    - email
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.LoginRequest:
    properties:
      cpf:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - cpf
    - password
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.PhoneRequest:
    properties:
      phone:
        type: string
      principal:
        type: boolean
    required:
    - phone
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.RefreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.SignupRequest:
    properties:
      cpf:
        type: string
      device:
        $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.Device'
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - cpf
    - device
    - email
    - name
    - password
    type: object
  github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.UpdateUserRequest:
    properties:
      addresses:
        items:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.AddressRequest'
        type: array
      emails:
        items:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.EmailRequest'
        type: array
      emergency_contact:
        type: string
      name:
        type: string
      phones:
        items:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.PhoneRequest'
        type: array
    type: object
host: api.pohinc.com.br
info:
  contact: {}
  description: API para gerenciamento de pacientes e dispositivos via Keycloak e polyglot
    persistence.
  title: Sistema Centralizador de Dados Clínicos API
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: Autentica um paciente via Keycloak exigindo DPoP proof (RFC 9449)
        no header
      parameters:
      - description: DPoP Proof JWT (RFC 9449)
        in: header
        name: DPoP
        required: true
        type: string
      - description: Credenciais do paciente
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login de Paciente com DPoP
      tags:
      - users
  /api/refresh:
    post:
      consumes:
      - application/json
      description: Emite um novo token usando um RefreshToken via Keycloak exigindo
        DPoP proof no header
      parameters:
      - description: DPoP Proof JWT (RFC 9449)
        in: header
        name: DPoP
        required: true
        type: string
      - description: Token de refresh
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualização de Access Token
      tags:
      - users
  /api/signup:
    post:
      consumes:
      - application/json
      description: Cadastra um paciente integrando Keycloak, MariaDB e Cassandra
      parameters:
      - description: Dados do paciente e dispositivo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cadastro de Paciente
      tags:
      - users
  /api/users:
    put:
      consumes:
      - application/json
      description: Atualiza dados do usuário (nome, telefone, endereço) no MariaDB
        e Keycloak
      parameters:
      - description: Access Token (Bearer or DPoP)
        in: header
        name: Authorization
        required: true
        type: string
      - description: DPoP Proof JWT (RFC 9449)
        in: header
        name: DPoP
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TCC-Conjunto-de-Aplicacoes-Medicinais_sistema_centralizador_de_dados_clinicos_back_shared_models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualização de Usuário
      tags:
      - users
  /api/users/{id}/send-verify-email:
    post:
      description: Solicita ao Keycloak o envio de um e-mail de verificação para o
        usuário
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Enviar E-mail de Verificação
      tags:
      - users
swagger: "2.0"